Bullet firing system has been implemented succesfully. Suggestions has been also implemented in the system.
