+++ Assembly-CSharp.csproj
@@ -67,15 +67,23 @@
     <Compile Include="Assets\MyAssets\Scripts\MVC\Bullet\BulletModel.cs" />
     <Compile Include="Assets\MyAssets\Scripts\MVC\Bullet\BulletView.cs" />
     <Compile Include="Assets\MyAssets\Scripts\MVC\Bullet\Bullet_Service.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\EnemyTanks\EnemyTankService.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\EnemyTanks\EnemyView.cs" />
     <Compile Include="Assets\MyAssets\Scripts\MVC\Enums\BulletType.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\Enums\EnemyType.cs" />
     <Compile Include="Assets\MyAssets\Scripts\MVC\Enums\TankType.cs" />
-    <Compile Include="Assets\MyAssets\Scripts\MVC\ScriptablrObjects\TankScriptableObjectList.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\PlayerTanks\TankController.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\PlayerTanks\TankModel.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\PlayerTanks\TankService.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\PlayerTanks\TankView.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\ScriptablrObjects\BulletScriptables\BulletScriptableObject.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\ScriptablrObjects\BulletScriptables\BulletScriptableObjectList.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\ScriptablrObjects\EnemyScriptables\EnemyScriptableObject.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\ScriptablrObjects\EnemyScriptables\EnemyScriptableObjectList.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\ScriptablrObjects\TankScriptables\TankScriptableObject.cs" />
+    <Compile Include="Assets\MyAssets\Scripts\MVC\ScriptablrObjects\TankScriptables\TankScriptableObjectList.cs" />
     <Compile Include="Assets\MyAssets\Scripts\MVC\Singleton\MonoSingleton.cs" />
     <Compile Include="Assets\MyAssets\Scripts\MVC\Singleton\MonoSingletonGeneric.cs" />
-    <Compile Include="Assets\MyAssets\Scripts\MVC\Tank\TankController.cs" />
-    <Compile Include="Assets\MyAssets\Scripts\MVC\Tank\TankModel.cs" />
-    <Compile Include="Assets\MyAssets\Scripts\MVC\Tank\TankService.cs" />
-    <Compile Include="Assets\MyAssets\Scripts\MVC\Tank\TankView.cs" />
     <None Include="Assets\Fonts\Bowlby_One\OFL.txt" />
     <Reference Include="Unity.Timeline.Editor">
       <HintPath>C:/Users/Abhishek and Nitesh/Documents/battle-tank-game/Library/ScriptAssemblies/Unity.Timeline.Editor.dll</HintPath>
