+++ Assets/MyAssets/Scripts/MVC/Bullet/Shell.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 8911956426050616456}
   - component: {fileID: 8911956426054040744}
   - component: {fileID: 8911956426052910696}
+  - component: {fileID: 1610125316115076819}
   - component: {fileID: 434162905568129984}
   m_Layer: 0
   m_Name: Shell
@@ -80,6 +81,22 @@ MeshRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 0
+--- !u!54 &1610125316115076819
+Rigidbody:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8911956426050904232}
+  serializedVersion: 2
+  m_Mass: 1
+  m_Drag: 0
+  m_AngularDrag: 0.05
+  m_UseGravity: 1
+  m_IsKinematic: 0
+  m_Interpolate: 0
+  m_Constraints: 0
+  m_CollisionDetection: 0
 --- !u!114 &434162905568129984
 MonoBehaviour:
   m_ObjectHideFlags: 0
