+++ Assets/MyAssets/Scripts/MVC/EnemyTanks.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 130c40de2b622804d95c973540f64c36
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
