+++ Assets/MyAssets/Scripts/MVC/EnemyTanks/EnemyTank.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 07fecf7dd1397ea4a8ef0069578fe286
-NativeFormatImporter:
+guid: aa30cbfb03455b94e94a9dfd3e151db6
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
