+++ Assets/MyAssets/Scripts/MVC/EnemyTanks/EnemyTankService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 125dd2685c468a04996423656f26899d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
