+++ Assets/MyAssets/Scripts/MVC/EnemyTanks/EnemyView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b4ca7453a0f39a64eb8993cd0541dd33
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
