+++ Assets/MyAssets/Scripts/MVC/Enums/EnemyType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a3537662a8eb06a4b83828a8d51832da
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
