+++ Assets/MyAssets/Scripts/MVC/PlayerTanks/TankView.cs.meta
