+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c9e8f23215ef34e409f79767d6280c32
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
