+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables/BulletScriptableObjectList.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 0}
+  m_Name: BulletScriptableObjectList
+  m_EditorClassIdentifier: Assembly-CSharp:ScriptableObj:BulletScriptableObjectList
+  bullet:
+  - {fileID: 11400000, guid: 3c321e367b9fffd47b274226f87721af, type: 2}
+  - {fileID: 11400000, guid: a6618d36f0da8084c871578dc31c0c0d, type: 2}
+  - {fileID: 11400000, guid: edff7db3d6f68fa4399fd405d03578c5, type: 2}
