+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables/BulletScriptableObjectList.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: acef751b1b6b3ac4a8547d857931b89a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
