+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables/BulletScriptableObjectList.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Bullet;
+using Bullet.View;
+
+
+namespace ScriptableObj
+{
+    [CreateAssetMenu(fileName = "BulletScriptableObjectList", menuName = "ScriptableObjects/NewBulletListScriptableObject")]
+    public class BulletScriptableObjectList : ScriptableObject
+    {
+        public BulletScriptableObject[] bullet;
+    }
+}
