+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables/BulletScriptableObjectList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 691b4c76055a25248a2de9451c6e18a3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
