+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables/FastBullet.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 48821e78118b7ad499a4dfce864d7829, type: 3}
+  m_Name: FastBullet
+  m_EditorClassIdentifier: 
+  bulletType: 3
+  speed: 20
+  damage: 3
+  BulletView: {fileID: 434162905568129984, guid: 5169d41190353d54ba1618048a3fb9d4,
+    type: 3}
