+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables/FastBullet.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3c321e367b9fffd47b274226f87721af
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
