+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables/MediumBullet.asset
@@ -9,11 +9,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: RedTank
-  m_EditorClassIdentifier: Assembly-CSharp:Scriptables:TankScriptableObject
-  tankType: 1
-  tankName: RedTank
-  speed: 75
-  turn: 30
-  health: 200
+  m_Script: {fileID: 11500000, guid: 48821e78118b7ad499a4dfce864d7829, type: 3}
+  m_Name: MediumBullet
+  m_EditorClassIdentifier: 
+  bulletType: 2
+  speed: 15
+  damage: 5
+  BulletView: {fileID: 434162905568129984, guid: 5169d41190353d54ba1618048a3fb9d4,
+    type: 3}
