+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables/MediumBullet.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a6618d36f0da8084c871578dc31c0c0d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
