+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables/SlowBullet.asset
@@ -9,11 +9,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: BlueTank
-  m_EditorClassIdentifier: Assembly-CSharp:Scriptables:TankScriptableObject
-  tankType: 2
-  tankName: BlueTank
-  speed: 80
-  turn: 40
-  health: 100
+  m_Script: {fileID: 11500000, guid: 48821e78118b7ad499a4dfce864d7829, type: 3}
+  m_Name: SlowBullet
+  m_EditorClassIdentifier: 
+  bulletType: 1
+  speed: 10
+  damage: 10
+  BulletView: {fileID: 434162905568129984, guid: 5169d41190353d54ba1618048a3fb9d4,
+    type: 3}
