+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/BulletScriptables/SlowBullet.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: edff7db3d6f68fa4399fd405d03578c5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
