+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/EnemyScriptables.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 2dd5437d030b902428c5c4477e746740
-NativeFormatImporter:
+guid: daebbcb5a6fd1c944ab37a6b7620028f
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
