+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/EnemyScriptables/EnemyScriptableObjectList.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Enemy.View;
+
+namespace enemyScrtiptables
+{
+    [CreateAssetMenu(fileName = "EnemyTankScriptableObjectList", menuName = "ScriptableObjects/NewEnemyListScriptableObject")]
+    public class EnemyScriptableObjectList : ScriptableObject
+    {
+        public EnemyScriptableObject[] enemyTanks;
+    }
+}
