+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/EnemyScriptables/EnemyScriptableObjectList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b1b016c21b8c8ec46ab5740520e71f92
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
