+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/EnemyScriptables/EnemyTankScriptableObjectList.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b1b016c21b8c8ec46ab5740520e71f92, type: 3}
+  m_Name: EnemyTankScriptableObjectList
+  m_EditorClassIdentifier: 
+  enemyTanks:
+  - {fileID: 11400000, guid: da12ad709b849bf44b6c6e4e15e9fc09, type: 2}
+  - {fileID: 11400000, guid: 55fee651b09f9fe48abfcd5bac09b793, type: 2}
+  - {fileID: 11400000, guid: b3a33a12aff2ba64a808395d001a92fa, type: 2}
