+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/EnemyScriptables/EnemyTankScriptableObjectList.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2a0a102f88990d24488a86b2775c7913
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
