+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/EnemyScriptables/Heavy.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 05f7461d8664908438a58641c8fb3a48, type: 3}
+  m_Name: Heavy
+  m_EditorClassIdentifier: 
+  enemyType: 3
+  speed: 20
+  turn: 50
+  health: 0
+  BulletType: {fileID: 11400000, guid: edff7db3d6f68fa4399fd405d03578c5, type: 2}
+  enemyView: {fileID: 0}
