+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/EnemyScriptables/Heavy.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b3a33a12aff2ba64a808395d001a92fa
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
