+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/EnemyScriptables/Light.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: da12ad709b849bf44b6c6e4e15e9fc09
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
