+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/EnemyScriptables/Medium.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 05f7461d8664908438a58641c8fb3a48, type: 3}
+  m_Name: Medium
+  m_EditorClassIdentifier: 
+  enemyType: 0
+  speed: 25
+  turn: 55
+  health: 0
+  BulletType: {fileID: 11400000, guid: a6618d36f0da8084c871578dc31c0c0d, type: 2}
+  enemyView: {fileID: 0}
