+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/EnemyScriptables/Medium.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 55fee651b09f9fe48abfcd5bac09b793
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
