--- Assets/MyAssets/Scripts/MVC/ScriptablrObjects/PlayerTanks.asset
@@ -1,18 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 990fd76fde9c2364091c51d870d31e50, type: 3}
-  m_Name: PlayerTanks
-  m_EditorClassIdentifier: 
-  tanks:
-  - {fileID: 11400000, guid: 2f3a7e5004fc6c54da252d78415f289c, type: 2}
-  - {fileID: 11400000, guid: eae9c4cfdacf5414db1908837de180e2, type: 2}
-  - {fileID: 11400000, guid: aebf3f926b999ae48ae8fc78e98fa110, type: 2}
