+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 99dbe686c5aaa494aad88ebedeeb30f3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
