+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables/BlueTank.asset
@@ -1,0 +1,23 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 0}
+  m_Name: BlueTank
+  m_EditorClassIdentifier: Assembly-CSharp:Scriptables:TankScriptableObject
+  tankType: 2
+  tankName: BlueTank
+  speed: 25
+  turn: 80
+  health: 100
+  damage: 7
+  BulletType: {fileID: 11400000, guid: a6618d36f0da8084c871578dc31c0c0d, type: 2}
+  TankView: {fileID: -3329480094707986513, guid: f8bd4a7afefc1884ab7d15ff7e22e170,
+    type: 3}
