+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables/BlueTank.asset.meta
