+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables/GreenTank.asset
@@ -14,6 +14,10 @@ MonoBehaviour:
   m_EditorClassIdentifier: Assembly-CSharp:Scriptables:TankScriptableObject
   tankType: 3
   tankName: GreenTank
-  speed: 100
+  speed: 15
   turn: 50
-  health: 100
+  health: 150
+  damage: 4
+  BulletType: {fileID: 11400000, guid: edff7db3d6f68fa4399fd405d03578c5, type: 2}
+  TankView: {fileID: -3329480094707986513, guid: f8bd4a7afefc1884ab7d15ff7e22e170,
+    type: 3}
