+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables/GreenTank.asset.meta
