+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables/RedTank.asset
@@ -1,0 +1,23 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 0}
+  m_Name: RedTank
+  m_EditorClassIdentifier: Assembly-CSharp:Scriptables:TankScriptableObject
+  tankType: 1
+  tankName: RedTank
+  speed: 20
+  turn: 50
+  health: 50
+  damage: 10
+  BulletType: {fileID: 11400000, guid: 3c321e367b9fffd47b274226f87721af, type: 2}
+  TankView: {fileID: -3329480094707986513, guid: f8bd4a7afefc1884ab7d15ff7e22e170,
+    type: 3}
