+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables/RedTank.asset.meta
