+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables/TankScriptableObjectList.asset
@@ -9,8 +9,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 990fd76fde9c2364091c51d870d31e50, type: 3}
-  m_Name: PlayersTankList
+  m_Script: {fileID: 11500000, guid: c064f71bdf1ee8f458270c045f328ef5, type: 3}
+  m_Name: TankScriptableObjectList
   m_EditorClassIdentifier: 
   tanks:
   - {fileID: 11400000, guid: 2f3a7e5004fc6c54da252d78415f289c, type: 2}
