+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables/TankScriptableObjectList.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f42ef81e9178a104baa5f1401190c450
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
