+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables/TankScriptableObjectList.cs
@@ -2,22 +2,9 @@
 using System.Collections.Generic;
 using UnityEngine;
 using Tank.View;
-using ScriptableObj;
 
 namespace Scriptables
 {
-    [CreateAssetMenu(fileName = "TankScriptableObject", menuName = "ScriptableObjects/NewTankScriptableObject")]
-    public class TankScriptableObject : ScriptableObject
-    {
-        public TankType tankType;
-        public string tankName;
-        public float speed;
-        public float turn;
-        public float health;
-        public BulletScriptableObject BulletType;
-        public TankView TankView;
-    }
-
     [CreateAssetMenu(fileName = "TankScriptableObjectList", menuName = "ScriptableObjects/NewTankListScriptableObject")]
     public class TankScriptableObjectList : ScriptableObject
     {
