+++ Assets/MyAssets/Scripts/MVC/ScriptablrObjects/TankScriptables/TankScriptableObjectList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c064f71bdf1ee8f458270c045f328ef5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
